completed implementation of 8-bit alpha channel pixmaps in libAfterImage; added function to hopefully determine the most efficient depth of alpha channel (8 or 1); added 8-bit alpha support to icon_t and mystyles; fixed bugs in MyStyles cloning of mask; added twm rubberband; fixed Audio (see AS ML)

+++ src/Wharf/Wharf.c
@@ -3105,6 +3105,6 @@ x_error_handler (Display * dpy, XErrorEvent * error)
   fprintf (stderr, "%s: X Error\n"
 	   "%*s  Request %d, Error %d %d, Type: %d\n",
 	   MyName,
-	   strlen (MyName), "", error->request_code, error->error_code, error->minor_code, error->type);
+	   (int)strlen (MyName), "", error->request_code, error->error_code, error->minor_code, error->type);
   return 0;
 }
